requirements.txt
flask
python-dotenv
schedule
markdown
google-genai
pydantic
openai
anthropic
flask-login
werkzeug
